requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
Automat==24.8.1
backports.zoneinfo==0.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
Django==4.2.18
dnspython==2.6.1
email-validator==2.2.0
filelock==3.16.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.3.0
packaging==24.2
parsel==1.9.1
pillow==10.4.0
Protego==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic-core==2.27.2
PyDispatcher==2.0.7
pyOpenSSL==25.0.0
pypng==0.20220715.0
qrcode==7.4.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
service-identity==24.2.0
sqlparse==0.5.3
tldextract==5.1.2
tomli==2.2.1
twisted==24.11.0
typing-extensions==4.12.2
urllib3==2.2.3
w3lib==2.2.1
zope.interface==7.2
